requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
cloudinary==1.21.1
confusable-homoglyphs==3.2.0
Django==3.0.8
django-bootstrap3==14.1.0
django-bootstrap4==2.2.0
django-registration==3.1
django-tinymce==3.0.2
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
